agent/requirements.txt
python-dotenv
pydantic-settings
websockets
paramiko
aiohttp
psutil
aiosnmp
pysnmplib
netifaces
gputil